Left to do:
when i do the automated calibration i want captured fiducial to appear in the list of fiducials in the manual calibration
I want to turn the beam off (create a button that disables the red circle)
delete an autofocus button from block selection. It is not needed there

I want to be able to set another mkm per pixel ratio explicitely

add zoom to the center or to the beam (when zoom to the beam check that roi isnt going outside the boundary)

When im zooming right now, beam shifts. i dont like